Refactor ShopApp to put the logic in a custom hook
